+++ packages/playground/src/index.tsx
@@ -26,13 +26,17 @@
 
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { ComponentsProvider } from '@looker/components'
-import { InlineInputTextDemo } from './InlineInputText/InlineInputTextDemo'
+import { Flex, ComponentsProvider } from '@looker/components'
+import { SelectContent } from './Select/SelectDemo'
+import { SelectMultiDemo } from './Select/SelectMultiDemo'
 
 const App: React.FC = () => {
   return (
     <ComponentsProvider>
-      <InlineInputTextDemo />
+      <Flex>
+        <SelectContent />
+        <SelectMultiDemo />
+      </Flex>
     </ComponentsProvider>
   )
 }